+++ PalestreGoGO/Web/Utils/GoogleAPIUtils.cs
@@ -16,12 +16,12 @@ namespace Web.Utils
             return string.Format(GOOGLE_MAPS_API_URL, key);
         }
 
-        public static string GetStaticMapUrl(string nomeMarker, float latitudine, float longitudine, string key)
+        public static string GetStaticMapUrl(string nomeMarker, double latitudine, double longitudine, string key)
         {
             return $"https://maps.googleapis.com/maps/api/staticmap?markers=size:mid%7Clabel:{WebUtility.UrlEncode(nomeMarker)}%7C{latitudine},{longitudine}&size=640x250&scale=2&maptype=roadmap&zoom=15&key={key}";
         }
 
-        public static string GetExternalMapUrl(float latitudine, float longitudine)
+        public static string GetExternalMapUrl(double latitudine, double longitudine)
         {
             return $"https://www.google.com/maps/search/?api=1&query={latitudine.ToString(CultureInfo.InvariantCulture)},{longitudine.ToString(CultureInfo.InvariantCulture)}";
         }
